Accessibility: Move the disposableFocus HoC to a new higher-order folder
