+++ components/higher-order/disposable-focus/index.js
